Shelly config: ensure server URL resolves to a local IP address
Issue found when putting in production.

If the WAN address of the server is specified, the report from the Shelly device
to the server will transmit through the router, and depending on its
configuration it might apply masquerading to the connection, which means the
server will see the router’s IP as source IP (which does not allows finding the
device on the network).