Allow extensions and views to set the same header
This is required for the logout view, since the session extension may do a
Set-Cookie to refresh the session and the logout view will do a Set-Cookie to
erase it. If both set the headers, the one set by the view will be used.

Previously, waiting >5 minutes on a page and trying to logout would cause
an AssertionError because of the header conflict. Added a non-regression test
for this scenario.