ORM: Support raw fields returns
Add a get_raw_fields method on Query that works like a regular select but
returns the raw database values of the specified fields (in fact, it returns the
raw SQLite cursor).
This is useful when retrieving a lot of values from the database that will not
make use of the DBObject infrastructure. Creating an object for each database
row takes time, and some conversions (like the datetime to float conversion done
on datetime.datetime values) are expensive.