History graph (take 1)
Implement a history graph to show the history of temperature and humidity.
The current implementation dynamically layouts the X (time) axis and allows
two Y axes, each showing any number of series.
The graph can be panned left and right, has control buttons, and a legend.
The major downside of the current implementation is that since both Y axes share
the same graph, the number of horizontal lines cannot be adjusted for each, and
is set to a fixed value (20). This means that the number of units between each
line can vary depending on the range of data and will probably not be a round
number (like 5, 10 etc).
This could be fixed by giving each Y axis its own horizontal lines, splitting
the graph in two (or three, etc) horizontal parts but keeping the common X axis.
To be tested later.